(feat) move huobi to new location
