+++ hummingbot/connector/exchange/huobi/huobi_auth.py
