+++ hummingbot/connector/exchange/huobi/huobi_in_flight_order.pxd
