+++ hummingbot/connector/exchange/huobi/huobi_market.pxd
