+++ hummingbot/connector/exchange/huobi/huobi_market.pyx
@@ -40,23 +40,23 @@ from hummingbot.core.utils.async_utils import (
     safe_gather,
 )
 from hummingbot.logger import HummingbotLogger
-from hummingbot.market.huobi.huobi_api_order_book_data_source import HuobiAPIOrderBookDataSource
-from hummingbot.market.huobi.huobi_api_user_stream_data_source import (
+from hummingbot.connector.exchange.huobi.huobi_api_order_book_data_source import HuobiAPIOrderBookDataSource
+from hummingbot.connector.exchange.huobi.huobi_api_user_stream_data_source import (
     HUOBI_SUBSCRIBE_TOPICS,
     HUOBI_ACCOUNT_UPDATE_TOPIC,
     HUOBI_ORDER_UPDATE_TOPIC
 )
-from hummingbot.market.huobi.huobi_auth import HuobiAuth
-from hummingbot.market.huobi.huobi_in_flight_order import HuobiInFlightOrder
-from hummingbot.market.huobi.huobi_order_book_tracker import HuobiOrderBookTracker
-from hummingbot.market.huobi.huobi_utils import (
+from hummingbot.connector.exchange.huobi.huobi_auth import HuobiAuth
+from hummingbot.connector.exchange.huobi.huobi_in_flight_order import HuobiInFlightOrder
+from hummingbot.connector.exchange.huobi.huobi_order_book_tracker import HuobiOrderBookTracker
+from hummingbot.connector.exchange.huobi.huobi_utils import (
     convert_to_exchange_trading_pair,
     convert_from_exchange_trading_pair)
 from hummingbot.connector.trading_rule cimport TradingRule
 from hummingbot.connector.exchange_base import ExchangeBase
 from hummingbot.market.market_base import (
     s_decimal_NaN)
-from hummingbot.market.huobi.huobi_user_stream_tracker import HuobiUserStreamTracker
+from hummingbot.connector.exchange.huobi.huobi_user_stream_tracker import HuobiUserStreamTracker
 from hummingbot.core.utils.tracking_nonce import get_tracking_nonce
 from hummingbot.core.utils.estimate_fee import estimate_fee
