+++ hummingbot/user/user_balances.py
@@ -1,7 +1,7 @@
 from hummingbot.connector.exchange.binance.binance_market import BinanceMarket
 from hummingbot.market.bittrex.bittrex_market import BittrexMarket
 from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_market import CoinbaseProMarket
-from hummingbot.market.huobi.huobi_market import HuobiMarket
+from hummingbot.connector.exchange.huobi.huobi_market import HuobiMarket
 from hummingbot.connector.exchange.kucoin.kucoin_market import KucoinMarket
 from hummingbot.connector.exchange.liquid.liquid_market import LiquidMarket
 from hummingbot.connector.exchange.kraken.kraken_market import KrakenMarket
